+++ src/native/util/os_linux.cpp
@@ -5,6 +5,8 @@
 
 #include <sys/syscall.h>
 #include <sys/types.h>
+#include <grp.h>
+#include <limits.h>
 
 namespace noobaa
 {
@@ -24,21 +26,33 @@ get_current_uid()
 const uid_t ThreadScope::orig_uid = getuid();
 const gid_t ThreadScope::orig_gid = getgid();
 
+const std::vector<gid_t> ThreadScope::orig_groups = [] {
+    std::vector<gid_t> groups(NGROUPS_MAX);
+    int r = getgroups(NGROUPS_MAX, &groups[0]);
+    groups.resize(r);
+    return groups;
+}();
+
 /**
  * set the effective uid/gid of the current thread using direct syscalls -
  * we have to bypass the libc wrappers because posix requires it to syncronize
  * uid & gid to all threads which is undesirable in our case.
+ * setgroups(0, NULL) - unsets the supplementary group IDs for the current thread
  */
 void
 ThreadScope::change_user()
 {
     if (_uid != orig_uid || _gid != orig_gid) {
         // must change gid first otherwise will fail on permission
+        setgroups(0, NULL);
         MUST_SYS(syscall(SYS_setresgid, -1, _gid, -1));
         MUST_SYS(syscall(SYS_setresuid, -1, _uid, -1));
     }
 }
 
+/**
+ * restores the effective uid/gid & supplementary_groups to the orig_uid/orig_gid/orig_groups
+ */
 void
 ThreadScope::restore_user()
 {
@@ -46,6 +60,7 @@ ThreadScope::restore_user()
         // must restore uid first otherwise will fail on permission
         MUST_SYS(syscall(SYS_setresuid, -1, orig_uid, -1));
         MUST_SYS(syscall(SYS_setresgid, -1, orig_gid, -1));
+        setgroups(ThreadScope::orig_groups.size(), &ThreadScope::orig_groups[0]);
     }
 }
