+++ src/test/unit_tests/index.js
@@ -58,7 +58,6 @@ require('./test_bucket_chunks_builder');
 require('./test_mirror_writer');
 require('./test_namespace_fs');
 require('./test_ns_list_objects');
-require('./test_bucketspace');
 require('./test_chunk_fs');
 require('./test_s3select');
