+++ src/test/unit_tests/test_bucketspace_versioning.js
@@ -1,18 +1,20 @@
 /* Copyright (C) 2020 NooBaa */
-/*eslint max-lines-per-function: ["error", 1000]*/
+/* eslint-disable max-lines-per-function */
 'use strict';
 
-
+const fs = require('fs');
+const path = require('path');
+const http = require('http');
 const mocha = require('mocha');
 const AWS = require('aws-sdk');
-const http = require('http');
 const assert = require('assert');
 const coretest = require('./coretest');
 const { rpc_client, EMAIL } = coretest;
 const fs_utils = require('../../util/fs_utils');
-const size_utils = require('../../util/size_utils');
-const path = require('path');
 const nb_native = require('../../util/nb_native');
+const size_utils = require('../../util/size_utils');
+const test_utils = require('../system_tests/test_utils');
+coretest.setup({});
 
 const MAC_PLATFORM = 'darwin';
 const XATTR_VERSION_ID = 'user.version_id';
@@ -61,10 +63,7 @@ mocha.describe('bucketspace namespace_fs - versioning', function() {
     const dir1_versions_path = path.join(full_path, dir1, '.versions/');
 
     mocha.before(async function() {
-        if (process.getgid() !== 0 || process.getuid() !== 0) {
-            coretest.log('No Root permissions found in env. Skipping test');
-            this.skip(); // eslint-disable-line no-invalid-this
-        }
+        if (test_utils.invalid_nsfs_root_permissions()) this.skip(); // eslint-disable-line no-invalid-this
         // create paths 
         await fs_utils.create_fresh_path(tmp_fs_root, 0o777);
         await fs_utils.create_fresh_path(full_path, 0o770);
@@ -1020,6 +1019,426 @@ mocha.describe('bucketspace namespace_fs - versioning', function() {
     });
 });
 
+mocha.describe('bucketspace namespace_fs - versioning', function() {
+    mocha.describe('List-objects', function() {
+        const nsr = 'noobaa-nsr';
+        const bucket_name = 'noobaa-bucket';
+        let tmp_fs_root = '/tmp/test_namespace_fs_list_objects';
+        if (process.platform === MAC_PLATFORM) {
+            tmp_fs_root = '/private/' + tmp_fs_root;
+        }
+        const bucket_path = '/bucket';
+        const full_path = tmp_fs_root + bucket_path;
+        const version_dir = '/.versions';
+        const full_path_version_dir = full_path + `${version_dir}`;
+        const dir1 = full_path + '/dir1';
+        const dir1_version_dir = dir1 + `${version_dir}`;
+        const dir2 = full_path + '/dir2';
+        const dir2_version_dir = dir2 + `${version_dir}`;
+        let s3_client;
+        let s3_admin;
+        const accounts = [];
+        const key = 'key';
+        const body = 'AAAA';
+        const version_key = 'version_key';
+        const version_body = 'A1A1A1A';
+
+        mocha.before(async function() {
+            if (test_utils.invalid_nsfs_root_permissions()) this.skip(); // eslint-disable-line no-invalid-this
+            // create paths
+            await fs_utils.create_fresh_path(tmp_fs_root, 0o777);
+            await fs_utils.create_fresh_path(full_path, 0o770);
+            await fs_utils.file_must_exist(full_path);
+            await fs_utils.create_fresh_path(full_path_version_dir, 0o770);
+            await fs_utils.file_must_exist(full_path_version_dir);
+            await fs_utils.create_fresh_path(dir1, 0o770);
+            await fs_utils.file_must_exist(dir1);
+            await fs_utils.create_fresh_path(dir1_version_dir, 0o770);
+            await fs_utils.file_must_exist(dir1_version_dir);
+            await fs_utils.create_fresh_path(dir2, 0o770);
+            await fs_utils.file_must_exist(dir2);
+            await fs_utils.create_fresh_path(dir2_version_dir, 0o770);
+            await fs_utils.file_must_exist(dir2_version_dir);
+            // export dir as a bucket
+            await rpc_client.pool.create_namespace_resource({
+                name: nsr,
+                nsfs_config: {
+                    fs_root_path: tmp_fs_root,
+                }
+            });
+            const obj_nsr = { resource: nsr, path: bucket_path };
+            await rpc_client.bucket.create_bucket({
+                name: bucket_name,
+                namespace: {
+                    read_resources: [obj_nsr],
+                    write_resource: obj_nsr
+                }
+            });
+            const policy = {
+                Version: '2012-10-17',
+                Statement: [{
+                    Sid: 'id-1',
+                    Effect: 'Allow',
+                    Principal: { AWS: "*" },
+                    Action: ['s3:*'],
+                    Resource: [`arn:aws:s3:::*`]
+                },
+            ]
+            };
+            // create accounts
+            let res = await generate_nsfs_account({ admin: true });
+            s3_admin = generate_s3_client(res.access_key, res.secret_key);
+            await s3_admin.putBucketPolicy({
+                Bucket: bucket_name,
+                Policy: JSON.stringify(policy)
+            }).promise();
+            // create nsfs account
+            res = await generate_nsfs_account();
+            s3_client = generate_s3_client(res.access_key, res.secret_key);
+            accounts.push(res.email);
+            await s3_client.putBucketVersioning({ Bucket: bucket_name, VersioningConfiguration: { MFADelete: 'Disabled', Status: 'Enabled' } }).promise();
+            const bucket_ver = await s3_client.getBucketVersioning({ Bucket: bucket_name }).promise();
+            assert.equal(bucket_ver.Status, 'Enabled');
+            await create_object(`${full_path}/${key}`, body, 'null');
+            await create_object(`${full_path_version_dir}/${version_key}`, version_body, 'null');
+            await create_object(`${dir1}/${key}`, body, 'null');
+            await create_object(`${dir1_version_dir}/${version_key}`, version_body, 'null');
+            await create_object(`${dir2}/${key}`, body, 'null');
+            await create_object(`${dir1_version_dir}/${version_key}`, version_body, 'null');
+        });
+
+        mocha.after(async () => {
+            fs_utils.folder_delete(tmp_fs_root);
+            for (const email of accounts) {
+                await rpc_client.account.delete_account({ email });
+            }
+        });
+
+        mocha.it('list objects - should return only latest object', async function() {
+            const res = await s3_client.listObjects({Bucket: bucket_name}).promise();
+            res.Contents.forEach(val => {
+                if (val.Key.includes('version') === true) {
+                    assert.fail('Not Expected: list objects returned contents fo .version dir');
+                }
+            });
+        });
+    });
+
+    mocha.describe('Get/Head object', function() {
+        const nsr = 'get-head-versioned-nsr';
+        const bucket_name = 'get-head-versioned-bucket';
+        const disabled_bucket_name = 'get-head-disabled-bucket';
+        let tmp_fs_root = '/tmp/test_namespace_fs_get_objects';
+        if (process.platform === MAC_PLATFORM) {
+            tmp_fs_root = '/private/' + tmp_fs_root;
+        }
+
+        const bucket_path = '/get-head-bucket';
+        const vesion_dir = '/.versions';
+        const full_path = tmp_fs_root + bucket_path;
+        const disabled_bucket_path = '/get-head-disabled_bucket';
+        const disabled_full_path = tmp_fs_root + disabled_bucket_path;
+        const version_dir_path = full_path + vesion_dir;
+        let file_pointer;
+        const versionID_1 = 'mtime-12a345b-ino-c123d45';
+        const versionID_2 = 'mtime-e56789f-ino-h56g789';
+        const versionID_3 = 'mtime-1i357k9-ino-13l57j9';
+        let s3_client;
+        let s3_admin;
+        const accounts = [];
+        const dis_version_key = 'dis_version';
+        const dis_version_body = 'AAAAA';
+        const en_version_key = 'en_version';
+        const en_version_body = 'BBBBB';
+        const en_version_key_v1 = versionID_2;
+        const en_version_body_v1 = 'CCCCC';
+        const key_version = en_version_key + '_' + en_version_key_v1;
+        mocha.before(async function() {
+            const self = this; // eslint-disable-line no-invalid-this
+            self.timeout(300000);
+            if (test_utils.invalid_nsfs_root_permissions()) this.skip(); // eslint-disable-line no-invalid-this
+            // create paths
+            await fs_utils.create_fresh_path(tmp_fs_root, 0o777);
+            await fs_utils.create_fresh_path(full_path, 0o770);
+            await fs_utils.file_must_exist(full_path);
+            await fs_utils.create_fresh_path(version_dir_path, 0o770);
+            await fs_utils.file_must_exist(version_dir_path);
+            await fs_utils.create_fresh_path(disabled_full_path, 0o770);
+            await fs_utils.file_must_exist(disabled_full_path);
+            // export dir as a bucket
+            await rpc_client.pool.create_namespace_resource({
+                name: nsr,
+                nsfs_config: {
+                    fs_root_path: tmp_fs_root,
+                }
+            });
+            const obj_nsr = { resource: nsr, path: bucket_path };
+            await rpc_client.bucket.create_bucket({
+                name: bucket_name,
+                namespace: {
+                    read_resources: [obj_nsr],
+                    write_resource: obj_nsr
+                }
+            });
+
+            const disabled_nsr = { resource: nsr, path: disabled_bucket_path };
+            await rpc_client.bucket.create_bucket({
+                name: disabled_bucket_name,
+                namespace: {
+                    read_resources: [disabled_nsr],
+                    write_resource: disabled_nsr
+                }
+            });
+
+            const policy = {
+                Version: '2012-10-17',
+                Statement: [{
+                    Sid: 'id-1',
+                    Effect: 'Allow',
+                    Principal: { AWS: "*" },
+                    Action: ['s3:*'],
+                    Resource: [`arn:aws:s3:::*`]
+                },
+            ]
+            };
+            // create accounts
+            let res = await generate_nsfs_account({ admin: true });
+            s3_admin = generate_s3_client(res.access_key, res.secret_key);
+            await s3_admin.putBucketPolicy({
+                Bucket: bucket_name,
+                Policy: JSON.stringify(policy)
+            }).promise();
+            await s3_admin.putBucketPolicy({
+                Bucket: disabled_bucket_name,
+                Policy: JSON.stringify(policy)
+            }).promise();
+            // create nsfs account
+            res = await generate_nsfs_account();
+            s3_client = generate_s3_client(res.access_key, res.secret_key);
+            accounts.push(res.email);
+            // create a file in version disabled bucket
+            await create_object(`${disabled_full_path}/${dis_version_key}`, dis_version_body, 'null');
+            // create a file after when versioning not enabled
+            await create_object(`${full_path}/${dis_version_key}`, dis_version_body, 'null');
+            await s3_client.putBucketVersioning({ Bucket: bucket_name, VersioningConfiguration: { MFADelete: 'Disabled', Status: 'Enabled' } }).promise();
+            const bucket_ver = await s3_client.getBucketVersioning({ Bucket: bucket_name }).promise();
+            assert.equal(bucket_ver.Status, 'Enabled');
+            // create base file after versioning enabled
+            file_pointer = await create_object(`${full_path}/${en_version_key}`, en_version_body, versionID_1, true);
+        });
+
+        mocha.after(async () => {
+            if (file_pointer) await file_pointer.close(DEFAULT_FS_CONFIG);
+            fs_utils.folder_delete(tmp_fs_root);
+            for (const email of accounts) {
+                await rpc_client.account.delete_account({ email });
+            }
+        });
+
+        mocha.it('get object, versioning not enabled - should return latest object', async function() {
+            const res = await s3_client.getObject({Bucket: disabled_bucket_name, Key: dis_version_key}).promise();
+            assert.equal(res.Body, dis_version_body);
+        });
+
+        mocha.it('get object, versioning not enabled, version id specified - should return NoSuchKey', async function() {
+            try {
+                await s3_client.getObject({Bucket: disabled_bucket_name, Key: dis_version_key, VersionId: versionID_1}).promise();
+                assert.fail('Should fail');
+            } catch (err) {
+                    assert.equal(err.code, 'NoSuchKey');
+            }
+        });
+
+        mocha.it('get object, file created before the version is enabled - should return latest object', async function() {
+            const res = await s3_client.getObject({Bucket: bucket_name, Key: dis_version_key}).promise();
+            assert.equal(res.Body, dis_version_body);
+        });
+
+        mocha.it('get object, file created before the version is enabled, version id specified - should return ENOENT', async function() {
+            try {
+                    await s3_client.getObject({Bucket: bucket_name, Key: dis_version_key, VersionId: versionID_1}).promise();
+                    assert.fail('Should fail');
+            } catch (err) {
+                    assert.equal(err.code, 'NoSuchKey');
+            }
+        });
+
+        mocha.it('get object, with version enabled, no version id specified - should return latest object', async function() {
+            const res = await s3_client.getObject({Bucket: bucket_name, Key: en_version_key}).promise();
+            assert.equal(res.Body, en_version_body);
+        });
+
+        mocha.it('get object, with version enabled, with version id specified and matching with main file - should return latest object', async function() {
+            const res = await s3_client.getObject({Bucket: bucket_name, Key: en_version_key, VersionId: versionID_1}).promise();
+            assert.equal(res.Body, en_version_body);
+        });
+
+        mocha.it('get object, with version enabled, with version id specified and not matching with main file - should return $VersionId object', async function() {
+            await create_object(`${version_dir_path}/${key_version}`, en_version_body_v1, versionID_2);
+            const res = await s3_client.getObject({Bucket: bucket_name, Key: en_version_key, VersionId: versionID_2}).promise();
+            assert.equal(res.Body, en_version_body_v1);
+        });
+
+        mocha.it('get object, with version enabled, with version id specified and versioned object not present - should return NoSuchKey', async function() {
+            try {
+                await s3_client.getObject({Bucket: bucket_name, Key: en_version_key, VersionId: versionID_3}).promise();
+                assert.fail('Should fail');
+            } catch (err) {
+                assert.equal(err.code, 'NoSuchKey');
+            }
+        });
+
+        mocha.it('get object, with version enabled, with wrong version id specified - should return Bad Request', async function() {
+            try {
+                await s3_client.getObject({Bucket: bucket_name, Key: en_version_key, VersionId: 'ctime-12-ino-12a'}).promise();
+                assert.fail('Should fail');
+            } catch (err) {
+                assert.equal(err.code, 'BadRequest');
+            }
+        });
+
+        mocha.it('head object, version not enabled - should return latest object md', async function() {
+            try {
+                await s3_client.headObject({Bucket: disabled_bucket_name, Key: dis_version_key}).promise();
+                assert.ok('Expected latest head object returned');
+            } catch (err) {
+                assert.fail('Latest head object not found');
+            }
+        });
+
+        mocha.it('head object, version not enabled, version specified - should return object NotFound', async function() {
+            try {
+                await s3_client.headObject({Bucket: disabled_bucket_name, Key: dis_version_key, VersionId: versionID_1}).promise();
+                assert.fail('Should fail');
+            } catch (err) {
+                assert.equal(err.code, 'NotFound');
+            }
+        });
+
+        mocha.it('head object, file created before the version is enabled - should return latest object md', async function() {
+            try {
+                await s3_client.headObject({Bucket: bucket_name, Key: dis_version_key}).promise();
+                assert.ok('Expected latest head object returned');
+            } catch (err) {
+                assert.fail('Latest head object not found');
+            }
+        });
+
+        mocha.it('head object, file created before the version is enabled, version specified - should return NotFound', async function() {
+            try {
+                await s3_client.headObject({Bucket: bucket_name, Key: dis_version_key, VersionId: versionID_1}).promise();
+                assert.ok('Expected latest head object returned');
+            } catch (err) {
+                assert.equal(err.code, 'NotFound');
+            }
+        });
+
+        mocha.it('head object, with version enabled, no version id specified - should return latest object md', async function() {
+            try {
+                await s3_client.headObject({Bucket: bucket_name, Key: en_version_key}).promise();
+                assert.ok('Expected latest head object returned');
+            } catch (err) {
+                assert.fail(`Failed with an error: ${err.code}`);
+            }
+        });
+
+        mocha.it('head object, with version enabled, with version id specified and matching with main file - should return latest object md', async function() {
+            try {
+                await s3_client.headObject({Bucket: bucket_name, Key: en_version_key, VersionId: versionID_1}).promise();
+                assert.ok('Expected latest head object returned');
+            } catch (err) {
+                assert.fail(`Failed with an error: ${err.code}`);
+            }
+        });
+
+        mocha.it('head object, with version enabled, with version id specified and not matching with main file - should return $VersionId object md', async function() {
+            await create_object(`${version_dir_path}/${key_version}`, en_version_body_v1, versionID_2);
+            try {
+                await s3_client.headObject({Bucket: bucket_name, Key: en_version_key, VersionId: versionID_2}).promise();
+                assert.ok('Expected versioned object returned');
+            } catch (err) {
+                assert.fail(`Failed with an error: ${err.code}`);
+            }
+        });
+
+        mocha.it('head object, with version enabled, versioned object not created - should return NotFound', async function() {
+            try {
+                await s3_client.headObject({Bucket: bucket_name, Key: en_version_key, VersionId: versionID_3}).promise();
+                assert.fail('Should fail');
+            } catch (err) {
+                assert.equal(err.code, 'NotFound');
+            }
+        });
+
+        mocha.it('head object, with version enabled, wrong version id specified - should return NotFound', async function() {
+            try {
+                await s3_client.headObject({Bucket: bucket_name, Key: en_version_key, VersionId: versionID_3}).promise();
+                assert.fail('Should fail');
+            } catch (err) {
+                assert.equal(err.code, 'NotFound');
+            }
+        });
+        mocha.it('Put object when version disbled and do put on the same object when version enabled - get object should return versioned object', async function() {
+            let res = await s3_client.getObject({Bucket: disabled_bucket_name, Key: dis_version_key}).promise();
+            assert.equal(res.Body, dis_version_body);
+            const version_obj_path = disabled_full_path + '/.versions';
+            const version_key = dis_version_key + '_null';
+            const version_body = 'DDDDD';
+            const version_body_new_version = 'EEEEE';
+            await s3_client.putBucketVersioning({ Bucket: disabled_bucket_name, VersioningConfiguration: { MFADelete: 'Disabled', Status: 'Enabled' } }).promise();
+            const bucket_ver = await s3_client.getBucketVersioning({ Bucket: disabled_bucket_name }).promise();
+            assert.equal(bucket_ver.Status, 'Enabled');
+            // May be below 3 steps can be replaced with put_object() to create versioned object
+            await fs_utils.create_fresh_path(version_obj_path, 0o770);
+            await fs_utils.file_must_exist(version_obj_path);
+            await create_object(`${version_obj_path}/${version_key}`, version_body, 'null');
+            await create_object(`${disabled_full_path}/${dis_version_key}`, version_body_new_version, 'mtime-jhdfbkjsd-ino-bnsdf7f');
+            res = await s3_client.getObject({Bucket: disabled_bucket_name, Key: dis_version_key, VersionId: 'null'}).promise();
+            assert.equal(res.Body, version_body);
+        });
+
+        mocha.it('get object, with version enabled, delete marker placed on latest object - should return NoSuchKey', async function() {
+            const xattr_delete_marker = { 'user.delete_marker': 'true' };
+            await file_pointer.replacexattr(DEFAULT_FS_CONFIG, xattr_delete_marker);
+            try {
+                await s3_client.getObject({Bucket: bucket_name, Key: en_version_key}).promise();
+                assert.fail('Should fail');
+            } catch (err) {
+                assert.equal(err.code, 'NoSuchKey');
+            }
+        });
+
+        mocha.it('get object, with version enabled, delete marker placed on latest object, version id specified - should return the version object', async function() {
+            const xattr_delete_marker = { 'user.delete_marker': 'true' };
+            await file_pointer.replacexattr(DEFAULT_FS_CONFIG, xattr_delete_marker);
+            const res = await s3_client.getObject({Bucket: bucket_name, Key: en_version_key, VersionId: versionID_2}).promise();
+            assert.equal(res.Body, en_version_body_v1);
+        });
+
+        mocha.it('head object, with version enabled, delete marked xattr placed on latest object - should return ENOENT', async function() {
+            const xattr_delete_marker = { 'user.delete_marker': 'true' };
+            await file_pointer.replacexattr(DEFAULT_FS_CONFIG, xattr_delete_marker);
+            try {
+                await s3_client.headObject({Bucket: bucket_name, Key: en_version_key}).promise();
+                assert.fail('Should fail');
+            } catch (err) {
+                assert.equal(err.code, 'NotFound');
+            }
+        });
+
+        mocha.it('head object, with version enabled, delete marked xattr placed on latest object, version id specified - should return the version object', async function() {
+            const xattr_delete_marker = { 'user.delete_marker': 'true' };
+            await file_pointer.replacexattr(DEFAULT_FS_CONFIG, xattr_delete_marker);
+            try {
+                await s3_client.headObject({Bucket: bucket_name, Key: en_version_key, VersionId: versionID_2}).promise();
+                assert.ok('Expected versioned object returned');
+            } catch (err) {
+                assert.fail(`Failed with an error: ${err.code}`);
+            }
+        });
+    });
+});
+
 
 
 /////// UTILS ///////
@@ -1228,5 +1647,14 @@ async function generate_nsfs_account(options = {}) {
     };
 }
 
-exports.generate_nsfs_account = generate_nsfs_account;
-exports.generate_s3_client = generate_s3_client;
+async function create_object(object_path, data, version_id, return_fd) {
+    const target_file = await nb_native().fs.open(DEFAULT_FS_CONFIG, object_path, 'w+');
+    await fs.promises.writeFile(object_path, data);
+    if (version_id !== 'null') {
+        const xattr_version_id = { 'user.version_id': `${version_id}` };
+        await target_file.replacexattr(DEFAULT_FS_CONFIG, xattr_version_id);
+    }
+    if (return_fd) return target_file;
+    await target_file.close(DEFAULT_FS_CONFIG);
+}
+
